Update main.cpp
Added Lives for both CPU and the Player and modified the winning functionality such that when the CPU lives become zero the player wins. Also added a high score functionality which keeps track of the player high score and stores it in a file and updates it whenever a new high score is reached.